source/server/requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
cffi==1.15.0
importlib-metadata==4.10.0
keyring==23.5.0
keyrings.cryptfile==1.3.8
pycparser==2.21
pycryptodome==3.12.0
zipp==3.7.0
pyftpdlib